feat(svm): create types file
